Day-2/app/requirements.txt
flask
mysql-connector-python
flask-restx
flask-cors
